fix: use static_cast<size_t> in NoiseState size guard for correct signedness comparison
Agent-Logs-Url: https://github.com/zjones2142/Peer2Pear/sessions/bb25a145-69d0-4144-80c6-2943f3dbfa12

+++ Peer2Pear_QtProj/NoiseState.cpp
@@ -86,7 +86,7 @@ QByteArray NoiseState::decryptAndHash(const QByteArray& ciphertext) {
         nonce[i] = static_cast<unsigned char>((m_n >> (8 * i)) & 0xff);
     }
 
-    if (ciphertext.size() < static_cast<int>(crypto_aead_xchacha20poly1305_ietf_ABYTES))
+    if (static_cast<size_t>(ciphertext.size()) < crypto_aead_xchacha20poly1305_ietf_ABYTES)
         return {};
 
     QByteArray pt;
